% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{DAT}
\alias{DAT}
\title{DAT}
\format{An object of class \code{character} of length 1.}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/NT_033778.4?report=fasta&from=16558567&to=16565238&strand=true}
}
\usage{
DAT
}
\description{
DNA sequence of dopamine uptake inhibitor. Used in examples.
}
\keyword{datasets}
