% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protospacers.R
\name{find_FASTA}
\alias{find_FASTA}
\title{find_FASTA}
\usage{
find_FASTA(
  file_fasta,
  chr = 7,
  start = 117465784,
  end = 117466784,
  l = 20,
  p = "NGG"
)
}
\arguments{
\item{file_fasta}{path to a file FASTA (either compressed or not).}

\item{chr}{chromosome number.}

\item{start}{start of the DNA sequence to scan for protospacers.}

\item{end}{end of the DNA sequence to scan for protospacers.}

\item{l}{protospacer length.}

\item{p}{PAM sequence to match.}
}
\value{
a dataframe with columns:
\enumerate{
    \item  chr.
    \item  start.
    \item  end.
    \item  protospacer sequence.
    \item  PAM sequence (e.g. CGG as the sequence that matched the PAM sequence NGG).
    \item  Strand (+ or -).
}
}
\description{
Function to identify all candidate guide (protospacer) sequences in a FASTA file. Note that the genomic
coordinates are \strong{1-indexed} and \strong{fully closed}.
}
