% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protospacers.R
\name{find_proto}
\alias{find_proto}
\title{find_proto}
\usage{
find_proto(d_seq = "TGATCTACTAGAGACTACTAACGGGGATACATAG", l = 20, PAM = "NGG")
}
\arguments{
\item{d_seq}{input DNA sequence for genomic region from which to identify protospacer sequences (in capital letters).}

\item{l}{protospacer length.}

\item{PAM}{sequence to match.}
}
\value{
a dataframe with columns.
\enumerate{
    \item  start.
    \item  end.
    \item  protospacer sequence.
    \item  PAM sequence (e.g. CGG as the sequence that matched the PAM sequence NGG).
    \item  Strand (+ or -).
    }
}
\description{
Function to identify all candidate guide (protospacer) sequences from a given genomic region.
}
\details{
For reverse strand reverseComplement PAM, comolement d_seq, match after PAM
}
