% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{solut_p2}
\alias{solut_p2}
\title{solut_p2}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 5 columns.}
\usage{
solut_p2
}
\description{
Solution part2.
}
\keyword{datasets}
