% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protospacers.R
\name{find_proto}
\alias{find_proto}
\title{find_proto}
\usage{
find_proto(s, l, p)
}
\arguments{
\item{s}{input DNA sequence for genomic region from which to identify protospacer sequences (in capital letters).}

\item{l}{protospacer length.}

\item{p}{PAM sequence to match (in capital letters).}
}
\value{
a list of protospacers.
}
\description{
Function to identify all candidate guide (protospacer) sequences from a given genomic region.
}
